+++ packages/app/src/cli/commands/app/dev.ts
@@ -63,6 +63,13 @@ export default class Dev extends Command {
       description: 'Override the ngrok tunnel URL. Format: "https://my-tunnel-url:port"',
       env: 'SHOPIFY_FLAG_TUNNEL_URL',
     }),
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    'no-tunnel': Flags.boolean({
+      hidden: true,
+      description: 'Automatic creation of a tunnel is disabled. Service entry point will listen to localhost instead',
+      env: 'SHOPIFY_FLAG_NO_TUNNEL',
+      default: false,
+    }),
   }
 
   public async run(): Promise<void> {
@@ -82,6 +89,7 @@ export default class Dev extends Command {
       subscriptionProductUrl: flags['subscription-product-url'],
       checkoutCartUrl: flags['checkout-cart-url'],
       tunnelUrl: flags['tunnel-url'],
+      noTunnel: flags['no-tunnel'],
     })
   }
 }