+++ packages/app/src/cli/services/dev/select-app.test.ts
@@ -67,8 +67,8 @@ describe('createApp', () => {
     const variables = {
       org: 2,
       title: 'app-name',
-      appUrl: 'https://shopify.github.io/shopify-cli/help/start-app/',
-      redir: ['http://localhost:3456'],
+      appUrl: 'https://example.com',
+      redir: ['https://example.com/api/auth'],
       type: 'custom',
     }
 
@@ -144,8 +144,8 @@ describe('selectOrCreateApp', () => {
     const variables = {
       org: 1,
       title: 'app-name',
-      appUrl: 'https://shopify.github.io/shopify-cli/help/start-app/',
-      redir: ['http://localhost:3456'],
+      appUrl: 'https://example.com',
+      redir: ['https://example.com/api/auth'],
       type: 'undecided',
     }