+++ packages/app/src/cli/services/scaffold/extension.ts
@@ -109,7 +109,7 @@ async function uiExtensionInit({
         title: `Scaffold ${getExtensionOutputConfig(extensionType).humanKey} extension`,
         task: async (_, task) => {
           task.title = `Scaffolding ${getExtensionOutputConfig(extensionType).humanKey} extension...`
-          const stdin = yaml.encode({
+          const input = yaml.encode({
             extensions: [
               {
                 title: name,
@@ -140,7 +140,7 @@ async function uiExtensionInit({
                 next()
               },
             }),
-            stdin,
+            input,
           })
           task.title = `${getExtensionOutputConfig(extensionType).humanKey} extension scaffolded`
         },
@@ -159,6 +159,7 @@ export function getRuntimeDependencies({
     case 'checkout_ui_extension':
     case 'pos_ui_extension':
     case 'web_pixel_extension':
+    case 'customer_accounts_ui_extension':
     case 'checkout_post_purchase': {
       const dependencies: DependencyVersion[] = [{name: 'react', version: versions.react}]
       const rendererDependency = getUIExtensionRendererDependency(extensionType)
@@ -175,34 +176,26 @@ async function functionExtensionInit(options: FunctionExtensionInitOptions) {
   await file.inTemporaryDirectory(async (tmpDir) => {
     const templateDownloadDir = path.join(tmpDir, 'download')
 
-    const list = ui.newListr(
-      [
-        {
-          title: `Scaffolding ${getExtensionOutputConfig(options.extensionType).humanKey} extension...`,
-          task: async (_, task) => {
-            await file.mkdir(templateDownloadDir)
-            await git.downloadRepository({
-              repoUrl: url,
-              destination: templateDownloadDir,
-              shallow: true,
-              progressUpdater: (statusString: string) => {
-                const taskOutput = `Cloning template from ${url}:\n${statusString}`
-                task.output = taskOutput
-              },
-            })
-            const origin = path.join(templateDownloadDir, functionTemplatePath(options))
-            await template.recursiveDirectoryCopy(origin, options.extensionDirectory, options)
-            const configYamlPath = path.join(options.extensionDirectory, 'script.config.yml')
-            if (await file.exists(configYamlPath)) {
-              await file.remove(configYamlPath)
-            }
-            task.title = `${getExtensionOutputConfig(options.extensionType).humanKey} extension scaffolded`
-          },
-        },
-      ],
-      {rendererSilent: environment.local.isUnitTest()},
-    )
-    await list.run()
+    await ui.task({
+      title: `Scaffolding ${getExtensionOutputConfig(options.extensionType).humanKey} extension...`,
+      task: async () => {
+        await file.mkdir(templateDownloadDir)
+        await git.downloadRepository({
+          repoUrl: url,
+          destination: templateDownloadDir,
+          shallow: true,
+        })
+        const origin = path.join(templateDownloadDir, functionTemplatePath(options))
+        await template.recursiveDirectoryCopy(origin, options.extensionDirectory, options)
+        const configYamlPath = path.join(options.extensionDirectory, 'script.config.yml')
+        if (await file.exists(configYamlPath)) {
+          await file.remove(configYamlPath)
+        }
+        return {
+          successMessage: `${getExtensionOutputConfig(options.extensionType).humanKey} extension scaffolded`,
+        }
+      },
+    })
   })
 }