+++ packages/app/src/cli/utilities/extensions/configuration.ts
@@ -95,6 +95,7 @@ export async function getUIExtensionResourceURL(uiExtensionType: UIExtensionType
     case 'checkout_post_purchase':
     case 'pos_ui_extension':
     case 'web_pixel_extension':
+    case 'customer_accounts_ui_extension':
       // This is a temporary workaround to avoid Admin crash when dev'ing multiple extensions
       // Issue at shopify/web: https://github.com/Shopify/web/blob/main/app/components/Extensions/hooks/useResourceUrlQuery.ts#L15-L37
       return {url: 'invalid_url'}
@@ -109,6 +110,8 @@ export function getUIExtensionSurface(uiExtensionType: UIExtensionTypes) {
       return 'checkout'
     case 'checkout_post_purchase':
       return 'post_purchase'
+    case 'customer_accounts_ui_extension':
+      return 'customer_accounts'
     case 'pos_ui_extension':
       return 'pos'
     case 'product_subscription':