+++ packages/cli-kit/src/node/ruby.ts
@@ -7,36 +7,43 @@ import constants from '../constants.js'
 import {coerce} from '../semver.js'
 import {AdminSession} from '../session.js'
 import {content, token} from '../output.js'
-// eslint-disable-next-line no-restricted-imports
-import {spawn} from 'child_process'
 import {Writable} from 'node:stream'
 
-const RubyCLIVersion = '2.16.0'
+const RubyCLIVersion = '2.20.1'
 const ThemeCheckVersion = '1.10.3'
 const MinBundlerVersion = '2.3.8'
 const MinRubyVersion = '2.3.0'
 const MinRubyGemVersion = '2.5.0'
 
+interface ExecCLI2Options {
+  // Contains token and store to pass to CLI 2.0, which will be set as environment variables
+  adminSession?: AdminSession
+  // Directory in which to execute the command. Otherwise the current directory will be used.
+  directory?: string
+}
 /**
  * Execute CLI 2.0 commands.
  * Installs a version of RubyCLI as a vendor dependency in a hidden folder in the system.
  * User must have a valid ruby+bundler environment to run any command.
  *
  * @param args {string[]} List of argumets to execute. (ex: ['theme', 'pull'])
- * @param adminSession {AdminSession} Contains token and store to pass to CLI 2.0, which will be set as environment variables
+ * @param options {ExecCLI2Options}
  */
-export async function execCLI2(args: string[], adminSession?: AdminSession) {
+export async function execCLI2(args: string[], {adminSession, directory}: ExecCLI2Options = {}) {
   await installCLIDependencies()
   const env = {
     ...process.env,
     SHOPIFY_CLI_ADMIN_AUTH_TOKEN: adminSession?.token,
     SHOPIFY_CLI_STORE: adminSession?.storeFqdn,
+    // Bundler uses this Gemfile to understand which gems are available in the
+    // environment. We use this to specify our own Gemfile for CLI2, which exists
+    // outside the user's project directory.
+    BUNDLE_GEMFILE: join(shopifyCLIDirectory(), 'Gemfile'),
   }
 
-  spawn('bundle', ['exec', 'shopify'].concat(args), {
+  await system.exec('bundle', ['exec', 'shopify'].concat(args), {
     stdio: 'inherit',
-    cwd: shopifyCLIDirectory(),
-    shell: true,
+    cwd: directory ?? process.cwd(),
     env,
   })
 }