+++ packages/cli-kit/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @shopify/cli-kit
 
+## 3.3.3
+
+### Patch Changes
+
+- e89dfa81: • Update Ruby CLI to latest
+  • Fix: Await Ruby subprocess
+  • Add path flag to init for better development flow
+  • Add verbose flag to see more output from the Node CLI
+  • Set up logging and CLI-kit store separately for the theme CLI
+
+## 3.3.2
+
+## 3.3.1
+
+## 3.3.0
+
+### Minor Changes
+
+- f7708fcc: Replace fastify/http-proxy with node-http-proxy to support Node < 17
+
 ## 3.2.0
 
 ### Patch Changes